added Experiment endpoints: create, get, delete

+++ src/main/java/com/seniorproject/first/prototype/service/ExperimentServiceImpl.java
@@ -6,7 +6,11 @@ import com.seniorproject.first.prototype.entity.User;
 import com.seniorproject.first.prototype.repository.ExperimentRepository;
 import com.seniorproject.first.prototype.repository.ParticipationRepository;
 import com.seniorproject.first.prototype.repository.UserRepository;
+import com.seniorproject.first.prototype.util.ResponseHandler;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -14,35 +18,141 @@ import org.springframework.stereotype.Service;
 import java.util.*;
 
 @Service
+@Slf4j
 public class ExperimentServiceImpl implements ExperimentService{
 
-    @Autowired
-    private ExperimentRepository experimentRepository;
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private ParticipationRepository participationRepository;
-    @Override
-    public Experiment createExperiment(Experiment experiment) {
+    private final ExperimentRepository experimentRepository;
+    private final UserRepository userRepository;
+    private final ParticipationRepository participationRepository;
 
-        List<Integer> overallResults = new ArrayList<>();
-        for(int i = 0; i < experiment.getWords().size(); i++){
-            overallResults.add(0);
-        }
-        experiment.setOverallResults(overallResults);
-        experiment.setParticipantCount((long)0);
+    @Autowired
+    public ExperimentServiceImpl(ExperimentRepository experimentRepository, UserRepository userRepository, ParticipationRepository participationRepository) {
+        this.experimentRepository = experimentRepository;
+        this.userRepository = userRepository;
+        this.participationRepository = participationRepository;
+    }
+
+    @Override
+    public ResponseEntity<Object> createExperiment(Experiment experiment) {
+        if(experiment.getWords() == null)
+            return ResponseHandler.generateResponse("Words are not provided", HttpStatus.BAD_REQUEST, null);
+
+        experiment.setFrequencyRange(null);
+        experiment.setNumberOfWords(experiment.getWords().size());
+        experiment.setLengthOfWords(null);
+        setOverallResults(experiment);
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String currentUserName = authentication.getName();
         User user = userRepository.findUserByUserEmail(currentUserName).get();
         experiment.setCreator(user);
 
-        return experimentRepository.save(experiment);
+        experimentRepository.save(experiment);
+
+        return ResponseHandler.generateResponse("Experiment " + experiment.getExperimentId() +" is created", HttpStatus.OK, experiment);
+
     }
 
     @Override
-    public Experiment getMyCreatedExperimentById(Long experimentId) {
-        return experimentRepository.findByExperimentId(experimentId);
+    public ResponseEntity<Object> createRandomizedExperiment(Experiment experiment) {
+        try {
+            log.info(experiment.toString());
+            log.info("Creating an experiment with randomized list of words at saveRandomizedExperimentV2()");
+            if(experiment.getNumberOfWords() == null)
+                return ResponseHandler.generateResponse("Could not create randomized list of words because number of words is not set", HttpStatus.BAD_REQUEST, null);
+            experiment.setWords(experimentRepository.findRandomWords(experiment.getNumberOfWords()));
+            experiment.setFrequencyRange(null);
+            experiment.setLengthOfWords(null);
+
+            setOverallResults(experiment);
+
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String currentUserName = authentication.getName();
+            User user = userRepository.findUserByUserEmail(currentUserName).get();
+            experiment.setCreator(user);
+
+            experimentRepository.save(experiment);
+        } catch (Exception e) {
+            log.error("Error at saveRandomizedExperiment(): ", e.getMessage());
+            return ResponseHandler.generateResponse("Could not create randomized list of words", HttpStatus.SERVICE_UNAVAILABLE, null);
+        }
+        log.info("Experiment " + experiment.getExperimentId() + " with randomized list of words is created");
+        return ResponseHandler.generateResponse("Experiment " + experiment.getExperimentId() +" with randomized list of words is created", HttpStatus.OK, experiment);
+
+    }
+
+    @Override
+    public ResponseEntity<Object> createExperimentLengthAndFrequency(Experiment experiment) {
+        if (experiment.getNumberOfWords() == null)
+            return ResponseHandler.generateResponse("Number of words is not set", HttpStatus.BAD_REQUEST, null);
+        if (experiment.getFrequencyRange() == null || experiment.getLengthOfWords() == null)
+            return ResponseHandler.generateResponse("Frequency range or word length is null", HttpStatus.BAD_REQUEST, null);
+        if (experiment.getFrequencyRange().size() != 2)
+            return ResponseHandler.generateResponse("Frequency range is invalid", HttpStatus.BAD_REQUEST, null);
+
+        Collections.sort(experiment.getFrequencyRange());
+        List<String> words = new ArrayList<>();
+        switch (experiment.getLengthOfWords()){
+            case 3:
+                words = experimentRepository.findWordsByFrequencyRangeAndLength3(experiment.getNumberOfWords(),
+                        experiment.getFrequencyRange().get(0), experiment.getFrequencyRange().get(1));
+                break;
+            case 4:
+                words = experimentRepository.findWordsByFrequencyRangeAndLength4(experiment.getNumberOfWords(),
+                        experiment.getFrequencyRange().get(0), experiment.getFrequencyRange().get(1));
+                break;
+            case 5:
+                words = experimentRepository.findWordsByFrequencyRangeAndLength5(experiment.getNumberOfWords(),
+                        experiment.getFrequencyRange().get(0), experiment.getFrequencyRange().get(1));
+                break;
+            case 6:
+                words = experimentRepository.findWordsByFrequencyRangeAndLength6(experiment.getNumberOfWords(),
+                        experiment.getFrequencyRange().get(0), experiment.getFrequencyRange().get(1));
+                break;
+            default:
+                return ResponseHandler.generateResponse("Cannot find words with length " + experiment.getLengthOfWords(), HttpStatus.BAD_REQUEST, null);
+        }
+
+        experiment.setWords(words);
+        setOverallResults(experiment);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserName = authentication.getName();
+        User user = userRepository.findUserByUserEmail(currentUserName).get();
+        experiment.setCreator(user);
+
+        experimentRepository.save(experiment);
+        return ResponseHandler.generateResponse("Experiment " + experiment.getExperimentId() +" is created", HttpStatus.OK, experiment);
+    }
+
+
+    @Override
+    public ResponseEntity<Object> getMyExperimentById(Long experimentId) {
+        Experiment experiment = experimentRepository.findByExperimentId(experimentId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if(experiment == null || !authentication.getName().equals(experiment.getCreator().getUserEmail()))
+            return ResponseHandler.generateResponse("Experiment with "+experimentId+" does not exist or belong to the user", HttpStatus.NOT_FOUND, null);
+
+        return ResponseHandler.generateResponse("Experiment " + experiment.getExperimentId() +" is returned", HttpStatus.OK, experiment);
+
+    }
+
+    @Override
+    public ResponseEntity<Object> getAllMyExperiments() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserName = authentication.getName();
+        User user = userRepository.findUserByUserEmail(currentUserName).get();
+
+        List<Experiment> experiments = experimentRepository.findAllByCreator(user);
+
+        return  ResponseHandler.generateResponse("Experiments are returned", HttpStatus.OK, experiments);
+    }
+
+    //TODO get all experiments in which the user has not participated and has not sent join request
+    @Override
+    public ResponseEntity<Object> getAllExperiments() {
+        return null;
     }
 
     @Override
@@ -88,15 +198,26 @@ public class ExperimentServiceImpl implements ExperimentService{
     }
 
     @Override
-    public void deleteMyCreatedExperimentById(Long experimentId) throws Exception {
+    public ResponseEntity<Object> deleteMyCreatedExperimentById(Long experimentId){
+        if(experimentRepository.findByExperimentId(experimentId) == null)
+            return ResponseHandler.generateResponse("Experiment " + experimentId +" is not found", HttpStatus.NOT_FOUND, experimentId);
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Experiment experiment = experimentRepository.findByExperimentId(experimentId);
-        if(!authentication.getName().equals(experiment.getCreator().getUserEmail())){
-            throw new Exception("Experiment was created by a different user");
-        }
-        else {
-            experimentRepository.deleteById(experimentId);
-        }
+        if(!authentication.getName().equals(experiment.getCreator().getUserEmail()))
+            return ResponseHandler.generateResponse("Experiment was created by a different user", HttpStatus.FORBIDDEN, experimentId);
+
+       try{
+           User user = userRepository.findUserByUserEmail(experiment.getCreator().getUserEmail()).get();
+           user.getCreatedExperiments().remove(experiment);
+           userRepository.save(user);
+           experimentRepository.deleteById(experimentId);
+       }
+       catch (Exception e){
+           return ResponseHandler.generateResponse(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, null);
+
+       }
+        return ResponseHandler.generateResponse("Experiment " + experimentId +" is deleted", HttpStatus.OK, experimentId);
+
     }
 
     @Override
@@ -125,8 +246,29 @@ public class ExperimentServiceImpl implements ExperimentService{
 
     //to be implemented
     @Override
-    public Experiment getMyCreatedExperimentByExperimentName(String experimentName) {
-        return experimentRepository.findByExperimentNameIgnoreCase(experimentName);
+    public ResponseEntity<Object> getMyCreatedExperimentByExperimentName(String experimentName) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserName = authentication.getName();
+        User user = userRepository.findUserByUserEmail(currentUserName).get();
+
+        List<Experiment> experiments = experimentRepository.findByExperimentNameIgnoreCaseAndCreator(experimentName, user);
+
+        if(experiments == null)
+            return ResponseHandler.generateResponse("Experiments with "+experimentName+" do not exist or belong to the user", HttpStatus.NOT_FOUND, null);
+
+        return ResponseHandler.generateResponse("Experiments are returned", HttpStatus.OK, experiments);
+
+    }
+
+
+
+    private void setOverallResults(Experiment experiment){
+        List<Integer> overallResults = new ArrayList<>();
+        for(int i = 0; i < experiment.getWords().size(); i++){
+            overallResults.add(0);
+        }
+        experiment.setOverallResults(overallResults);
+        experiment.setParticipantCount((long)0);
     }
 
 }
